+++ fdbclient/ClusterInterface.h
@@ -119,13 +119,14 @@ struct OpenDatabaseRequest {
 	Arena arena;
 	StringRef issues, traceLogGroup;
 	VectorRef<ClientVersionRef> supportedVersions;
+	int connectedCoordinatorsNum; // Number of coordinators connected by the client
 	UID knownClientInfoID;
 	ReplyPromise< struct ClientDBInfo > reply;
 
 	template <class Ar>
 	void serialize(Ar& ar) {
 		ASSERT( ar.protocolVersion() >= 0x0FDB00A400040001LL );
-		serializer(ar, issues, supportedVersions, traceLogGroup, knownClientInfoID, reply, arena);
+		serializer(ar, issues, supportedVersions, connectedCoordinatorsNum, traceLogGroup, knownClientInfoID, reply, arena);
 	}
 };