+++ fdbclient/Knobs.cpp
@@ -192,4 +192,7 @@ ClientKnobs::ClientKnobs(bool randomize) {
 	init( CONSISTENCY_CHECK_RATE_LIMIT_MAX,		  50e6 );
 	init( CONSISTENCY_CHECK_ONE_ROUND_TARGET_COMPLETION_TIME,	7 * 24 * 60 * 60 ); // 7 days
 	init( CONSISTENCY_CHECK_RATE_WINDOW,		  1.0  );
+
+	// TLS related
+	init( CHECK_CONNECTED_COORDINATOR_NUM_DELAY,  1.0 ); if( randomize && BUGGIFY ) CHECK_CONNECTED_COORDINATOR_NUM_DELAY =  g_random->random01() * 60.0; // In seconds
 }